backend/requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
postal==1.1.10
rapidfuzz==3.14.1
pandas==2.2.3
openpyxl==3.1.2
python-jose[cryptography]==3.5.0
httpx==0.27.0
sqlalchemy==2.0.29
psycopg[binary]>=3.3.0
alembic==1.17.2
python-dotenv==1.0.0
